services/tts-service/requirements-mlops.txt
# TTS service requirements
# For exact versions see constraints.txt

# Core numeric libraries (fundamental dependencies)
numpy
scipy

# Audio processing
librosa
soundfile
pydub

# ML monitoring
mlflow==2.8.0  # Pinned exact version to avoid backtracking
evidently==0.7.5  # Pinned to reduce resolver ambiguity
prometheus-client>=0.17.0  # For metrics

# Dashboard
streamlit==1.39.0  # Pinned to latest known working version
plotly>=5.18.0  # For data visualization
pandas>=2.0.0  # For data manipulation

# Feature store
redis>=4.5.5  # For feature store

# Cloud providers
boto3  # For AWS Polly and storage
azure-cognitiveservices-speech  # For Azure TTS
google-cloud-texttospeech  # For Google TTS

# gRPC for service communication
grpcio
grpcio-tools
protobuf

# Utilities
python-dotenv>=1.0.0
lxml>=4.9.0

# Language detection
langdetect==1.0.9

# Testing
pytest>=7.3.0
# Orchestration
apache-airflow==2.10.2  # For scheduling drift detection DAG
requests>=2.28.0  # For Slack webhooks and HTTP calls
# Training
scikit-learn>=1.2.0  # For model training
joblib>=1.2.0       # For model serialization
kafka-python>=2.0.2  # For Kafka streaming pipeline
flask>=2.2.0        # For HTTP feedback endpoint
torch>=1.13.0       # For user embedding encoder
torch-tensorrt>=1.4.0  # For TensorRT integration
onnx>=1.13.1           # For ONNX export
onnxruntime>=1.15.0    # For ONNX execution and quantization
onnxruntime-tools>=1.4.0  # For ONNX quantization utilities
tritonclient[all]>=2.14.0  # For Triton Inference Server client
